requirements.txt
agate==1.7.0
aiobotocore
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.2
anyio==3.6.2
apprise==1.3.0
asgi-lifespan==2.0.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Automat==22.10.0
Babel==2.12.1
beautifulsoup4==4.12.0
betterproto==1.2.5
boto3
botocore
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
constantly==15.1.0
coolname==2.2.0
croniter==1.3.8
cryptography==40.0.0
cssselect==1.2.0
dateparser==1.1.8
dbt-core==1.4.5
dbt-extractor==0.4.1
docker==6.0.1
exceptiongroup==1.1.1
fastapi==0.95.0
filelock==3.10.4
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
google-auth==2.16.3
graphql-core==3.2.3
greenlet==2.0.2
griffe==0.25.5
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hologram==0.0.15
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
isodate==0.6.1
itemadapter==0.7.0
itemloaders==1.0.6
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==3.2.0
kubernetes==26.1.0
leather==0.3.4
Logbook==1.5.3
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mashumaro==3.3.1
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
oauthlib==3.2.2
orjson==3.8.8
packaging==23.0
parsedatetime==2.4
parsel==1.7.0
pathspec==0.10.3
pendulum==2.1.2
pluggy==1.0.0
prefect==2.8.7
prefect-dbt==0.3.1
prefect-shell==0.1.5
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyDispatcher==2.0.7
Pygments==2.14.0
pyOpenSSL==23.1.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
queuelib==1.6.2
readchar==4.0.5
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
responses==0.23.1
rfc3986==1.5.0
rich==13.3.2
rsa==4.9
s3fs
s3transfer==0.6.0
Scrapy==2.8.0
service-identity==21.1.0
sgqlc==16.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
sqlparse==0.4.3
starlette==0.26.1
stringcase==1.2.0
text-unidecode==1.3
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
Twisted==22.10.0
typer==0.7.0
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
w3lib==2.1.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zope.interface==6.0
